Improve performance of gazelle --mode=diff
Don't invoke the "diff" command if the contents haven't changed.
Don't call Sync() on every write. Just closing the file suffices.

Before this change, diffing a repo with 1600 BUILD files would take 60 seconds.
After the change, it takes 2 seconds.

+++ cmd/gazelle/diff.go
@@ -16,28 +16,33 @@ limitations under the License.
 package main
 
 import (
+	"bytes"
 	"io/ioutil"
 	"os"
 	"os/exec"
 
-	bf "github.com/bazelbuild/buildtools/build"
 	"github.com/bazelbuild/bazel-gazelle/config"
+	bf "github.com/bazelbuild/buildtools/build"
 )
 
 func diffFile(c *config.Config, file *bf.File) error {
+	oldContents, err := ioutil.ReadFile(file.Path)
+	if err != nil {
+		oldContents = nil
+	}
+	newContents := bf.Format(file)
+	if bytes.Equal(oldContents, newContents) {
+		return nil
+	}
 	f, err := ioutil.TempFile("", c.DefaultBuildFileName())
 	if err != nil {
 		return err
 	}
+	f.Close()
 	defer os.Remove(f.Name())
-	defer f.Close()
-	if _, err := f.Write(bf.Format(file)); err != nil {
+	if err := ioutil.WriteFile(f.Name(), newContents, 0666); err != nil {
 		return err
 	}
-	if err := f.Sync(); err != nil {
-		return err
-	}
-
 	cmd := exec.Command("diff", "-u", "--new-file", file.Path, f.Name())
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
